Refactor InlineSearch and SearchResults for mobile responsiveness
- Replaced manual mobile detection in InlineSearch with a custom hook, simplifying the component and improving maintainability.
- Updated SearchResults to utilize the mobile detection hook, adjusting class names conditionally for better responsiveness.

These changes enhance the user experience by ensuring the components adapt seamlessly to different screen sizes, improving usability on mobile devices.
